==> Step 4 - Conducting Analysis.ipynb <==
# Step 4 - Conducting Analysis

This notebook takes in the data that was refined and outputted in Step 3, and runs analysis on it by producing several tables from that dataset. These tables contain several meaningful insights, which are observed and considered in the Readme.MD file.

#This cell defines constants and imports the notebook will use
import os, json
import pandas as pd

refined_data_path = "refined_data/wp_scored_city_articles_by_state.csv"

#This cell imports the outputted dataset from Step 3 and displays 
#it for a quick sanity check

wp_scored_city_articles_by_state = pd.read_csv(refined_data_path)
wp_scored_city_articles_by_state

##This cell defines the DataFrames that will be used to make tables
##for state-level analysis. It produces two DatFrames that will be used for 
##those tables, one that details how many city articles a state has overall,
#and one that specifies how many articles that ORES evaluated as high quality
#each state has. 

def is_high_quality(article_category):
    if article_category == "FA":
        return True;
    if article_category == "GA":
        return True;
    return False;
wp_scored_city_articles_by_state['is_high_quality_article'] = [is_high_quality(x) for x in list(wp_scored_city_articles_by_state['article_quality'])]
high_quality_subset = wp_scored_city_articles_by_state[wp_scored_city_articles_by_state['is_high_quality_article'] == True]

states_articles_agg = wp_scored_city_articles_by_state.groupby(['state', 'population'])['article_title'].agg('count')
states_articles_agg = states_articles_agg.reset_index(name = "article_count")

high_quality_subset_agg = high_quality_subset.groupby(['state', 'population'])['article_title'].agg('count')
high_quality_subset_agg = high_quality_subset_agg.reset_index(name = "hq_article_count")

states_articles_agg['article_count_over_population'] = states_articles_agg['article_count']/states_articles_agg['population']
high_quality_subset_agg['hq_article_count_over_population'] = high_quality_subset_agg['hq_article_count']/high_quality_subset_agg['population']

##This cell defines the DataFrames that will be used to make tables
##for regional division-level analysis. It produces two DatFrames that will be used for 
##those tables, one that details how many city articles a divion has overall,
#and one that specifies how many articles that ORES evaluated as high quality
#each division has. 

division_population_agg = pd.DataFrame({
    "state": wp_scored_city_articles_by_state['state'],
    "regional_division":  wp_scored_city_articles_by_state['regional_division'],
    "population": wp_scored_city_articles_by_state['population']
}).drop_duplicates() 
division_population_agg = division_population_agg.groupby(['regional_division'])['population'].agg('sum')
division_population_agg = division_population_agg.reset_index(name = "population")

division_aticles_agg = wp_scored_city_articles_by_state.groupby(['regional_division'])['article_title'].agg('count')
division_aticles_agg = division_aticles_agg.reset_index(name = "article_count")

division_agg = division_population_agg.join(division_aticles_agg.set_index('regional_division'), on='regional_division')
division_agg['article_count_over_population'] = division_agg['article_count']/division_agg['population']

hq_division_aticles_agg = high_quality_subset.groupby(['regional_division'])['article_title'].agg('count')
hq_division_aticles_agg = hq_division_aticles_agg.reset_index(name = "article_count")

hq_division_agg = division_population_agg.join(hq_division_aticles_agg.set_index('regional_division'), on='regional_division')
hq_division_agg['hq_article_count_over_population'] = hq_division_agg['article_count']/hq_division_agg['population']


## Analysis Tables

  

The below cells each display one table we will use for analyzing the data. Their implications and conclusions are further discussed in the Readme.MD file. The tables, in order, show the following:

  

1. The 10 US states with the highest total articles per capita (in descending order) .

2. The 10 US states with the lowest total articles per capita (in ascending order) .

3. The 10 US states with the highest high quality articles per capita (in descending order) .

4. The 10 US states with the lowest high quality articles per capita (in ascending order).

5. A rank ordered list of US census divisions (in descending order) by total articles per capita.

6. A rank ordered list of US census divisions (in descending order) by high quality articles per capita.

states_articles_agg.sort_values(by=['article_count_over_population'], ascending=False).head(10)

states_articles_agg.sort_values(by=['article_count_over_population'], ascending=True).head(10)

high_quality_subset_agg.sort_values(by=['hq_article_count_over_population'], ascending=False).head(10)

high_quality_subset_agg.sort_values(by=['hq_article_count_over_population'], ascending=True).head(10)

division_agg.sort_values(by=['article_count_over_population'], ascending=False)

hq_division_agg.sort_values(by=['hq_article_count_over_population'], ascending=False)
